requirements.txt
Flask==2.3.1
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
click==8.1.3
Werkzeug==2.3.2
python-dotenv==1.0.0
Bcrypt-Flask==1.0.2
WTForms==3.0.1
Flask-WTF==1.1.1
email-validator==2.0.0.post2